requirements.txt
requests
psycopg2-binary
tenacity
python-dotenv
pandas